+++ src/presenter/filters-presenter.js
@@ -2,7 +2,7 @@ import {
   render
 } from '../framework/render.js';
 
-import FiltersView from '../view/create-filter.js';
+import FiltersView from '../view/filters-view.js';
 
 const filtersContainer = document.querySelector('.trip-controls__filters');
