+++ src/view/create-trip-view.js
