+++ src/view/creator-point-view.js
