+++ src/view/editor-point-view.js
