+++ src/view/events-list-view.js
